Meer puntjes
bij printen onder de matches afwezigen neerzetten
plus bij positieve verandering

printen personenlijst alles behalve aanwezig met erboven "stand ladder na ronde " + vorigeronde

//resize matchmaking vorige week scores

gekko easter egg (up up down down left right left right b a) gekko plaatje

installeer applicatie met al ingevoerde personenlijst en associatie .d7p bestanden

+++ De 7 Pionnen/PersoonAanpassen.xaml.cs
@@ -60,7 +60,7 @@ namespace De_7_Pionnen
             geUpdatePersoon.Verloren = int.Parse(Verloren.Text);
             geUpdatePersoon.Gelijkspel = int.Parse(Gelijkspel.Text);
             geUpdatePersoon.glicko.Rating = float.Parse(Rating.Text.Replace('.', ','));
-            geUpdatePersoon.Score = float.Parse(Score.Text.Replace('.', ','));
+            geUpdatePersoon.Score = double.Parse(Score.Text.Replace('.', ','));
             
 
             Close();
